fix more float cast warnings

+++ src/widgets/updatebuttonbar.cpp
@@ -265,16 +265,16 @@ public:
 
 			if(flags & POPUPF_JUSTIFY_BUTTONS)
 			{
-				int left = 5;
+				double left = 5;
 				int count = 0;
 
-				int padding = ((GetWidth() - 10) - totalwidth)/btns.size();
+				double padding = ((GetWidth() - 10) - totalwidth)/btns.size();
 
 				for(auto &btn : btns)
 				{
 					count++;
 
-					int len = btn->GetPreferredWidth();
+					double len = btn->GetPreferredWidth();
 
 					if(count == btns.size())
 					{
@@ -290,7 +290,7 @@ public:
 			}
 			else
 			{
-				int left = (flags & POPUPF_CENTER_BUTTONS) ? (GetWidth() - (totalwidth + 5 * btns.size())) / 2 : 5;
+				double left = (flags & POPUPF_CENTER_BUTTONS) ? (GetWidth() - (totalwidth + 5 * btns.size())) / 2 : 5;
 				int count = 0;
 
 				bool ignore_right_align = (flags & POPUPF_CENTER_BUTTONS);
@@ -301,7 +301,7 @@ public:
 				{
 					bool align_right = !ignore_right_align && ((actions[count].flags & ACTIONF_FLOAT_RIGHT) || ((count + 1) == btns.size() && float_last_right));
 
-					int len = btn->GetPreferredWidth();
+					double len = btn->GetPreferredWidth();
 
 					if(!align_right)
 					{
@@ -315,8 +315,8 @@ public:
 
 				if(!ignore_right_align)
 				{
-					count = btns.size();
-					int right = GetWidth();
+					count = static_cast<int>(btns.size());
+					double right = GetWidth();
 
 					for(auto &btn : btns | std::views::reverse)
 					{
@@ -324,7 +324,7 @@ public:
 
 						bool align_right = (actions[count].flags & ACTIONF_FLOAT_RIGHT) || ((count + 1) == btns.size() && float_last_right);
 
-						int len = btn->GetPreferredWidth();
+						double len = btn->GetPreferredWidth();
 
 						if(align_right)
 						{
